Fix types to match C's more closely
Because the original libketama code uses floats, we need to make sure
we do the _exact_same_ floating point operations in Go in order to get the
same shard values.  A change in precision (from float32->float64 or vice
versa) changes the values just enough to assign different shards to
certain keys.  To maintain compatibility with the perl module, we need
the same floating point round-off errors.